requirements.txt
# Payer Portal Crawler Dependencies
# Healthcare Payer Knowledge Base System

# Web Scraping and Automation
selenium==4.15.0
beautifulsoup4==4.12.2
requests==2.31.0

# Web Driver Management
webdriver-manager==4.0.1

# Data Processing
pandas==2.1.3
numpy==1.25.2

# HTML/XML Processing
lxml==4.9.3

# Date/Time Processing
python-dateutil==2.8.2

# Logging and Configuration
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.23

# Data Validation
pydantic==2.5.0

# Progress Tracking
tqdm==4.66.1

# File Processing
pathlib2==2.3.7

# PDF Processing
PyPDF2==3.0.1
PyMuPDF==1.23.8

# Regular Expressions (enhanced)
regex==2023.10.3

# Network/HTTP handling
urllib3==2.1.0

# JSON Processing
ujson==5.8.0

# Parallel Processing
concurrent-futures==3.1.1

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0

# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Scheduler
APScheduler==3.10.4

# AI/ML - Embeddings and LLM
openai==1.3.0
anthropic==0.7.0
groq==0.33.0
sentence-transformers==2.2.2
torch==2.1.0

# Vector Operations
scikit-learn==1.3.2